api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
google-genai==1.11.0
langchain[anthropic,google-genai]
langchain-community
langchain-google-vertexai
langchain-core
langgraph
bs4
cognee